Refactoring shell scripts

+++ app/bin/wpphpunit.sh
@@ -1,18 +1,23 @@
 #!/usr/bin/env bash
 
-set -ex;
+set -e;
 
-dir=`dirname $0`
+themedir=$(wp theme path $(wp theme list --field=name --status=active) --dir)
+cd $themedir
 
-echo 'DROP DATABASE IF EXISTS wordpress_test;' | mysql -u root
+if [ -e $themedir/app/bin/install-wp-tests.sh ]; then
+  echo 'DROP DATABASE IF EXISTS wordpress_test;' | mysql -u root
 
-if [ -e /tmp/wordpress ]; then
-  rm -fr /tmp/wordpress
-fi
+  if [ -e /tmp/wordpress ]; then
+    rm -fr /tmp/wordpress
+  fi
 
-if [ -e /tmp/wordpress-tests-lib ]; then
-  rm -fr /tmp/wordpress-tests-lib
-fi
+  if [ -e /tmp/wordpress-tests-lib ]; then
+    rm -fr /tmp/wordpress-tests-lib
+  fi
 
-bash "$dir/install-wp-tests.sh" wordpress_test root '' localhost latest;
-phpunit
+  bash "$themedir/app/bin/install-wp-tests.sh" wordpress_test root '' localhost latest;
+  phpunit
+else
+  echo "$themedir/app/bin/install-wp-tests.sh not found."
+fi;
